+++ exporters/prometheus/prometheus.go
@@ -132,9 +132,9 @@ func (e *Exporter) Controller() *controller.Controller {
 	return e.controller
 }
 
-// ExportKindFor implements ExportKindSelector.
-func (e *Exporter) ExportKindFor(desc *sdkapi.Descriptor, kind aggregation.Kind) export.ExportKind {
-	return export.CumulativeExportKindSelector().ExportKindFor(desc, kind)
+// TemporalityFor implements TemporalitySelector.
+func (e *Exporter) TemporalityFor(desc *sdkapi.Descriptor, kind aggregation.Kind) aggregation.Temporality {
+	return aggregation.CumulativeTemporalitySelector().TemporalityFor(desc, kind)
 }
 
 // ServeHTTP implements http.Handler.